using second element in W_EVENT
