+++ data_prep/mapping/bufr_ascat.yaml
@@ -45,7 +45,7 @@ bufr:
 encoder:
   variables:
     # MetaData
-    - name: "MetaData/dateTime"
+    - name: "MetaData/time"
       source: variables/timestamp
       longName: "Observation Time"
       units: "seconds since 1970-01-01T00:00:00Z"
