+++ data_prep/scripts/gen_data.py
@@ -4,9 +4,20 @@ import argparse
 from datetime import datetime, timedelta
 
 base_path = os.path.split(os.path.realpath(__file__))[0]
-runner_path = os.path.realpath(os.path.join(base_path, '../src/create_data.py'))
 
 sys.path.append(os.path.realpath(os.path.join(base_path, '..', 'src')))
+
+
+def get_reader_path(data_type: str) -> str:
+    """
+    Returns the correct reader path based on the data_type.
+    """
+    if data_type == 'tank':
+        return os.path.realpath(os.path.join(base_path, '../src/tank_reader.py'))
+    elif data_type == 'bufr_pca':
+        return os.path.realpath(os.path.join(base_path, '../src/bufr_pca_reader.py'))
+    else:
+        raise ValueError(f"Unknown data_type: {data_type}")
 
 
 def _is_slurm_available() -> bool:
@@ -21,7 +32,8 @@ def _is_slurm_available() -> bool:
         return False
 
 
-def _make_sbatch_cmd(idx: int,
+def _make_sbatch_cmd(reader_path: str,
+                     idx: int,
                      start: datetime,
                      end: datetime,
                      ntasks: int,
@@ -44,7 +56,7 @@ def _make_sbatch_cmd(idx: int,
     cmd += f'--ntasks={ntasks} '
     cmd += '--time=02:00:00 '
     cmd += f'--job-name="gen_ocelot_{data_type}_{idx+1}" '
-    cmd += f'--wrap="srun -n {ntasks} python {runner_path} {start.strftime("%Y-%m-%d")} {end.strftime("%Y-%m-%d")} {data_type} '
+    cmd += f'--wrap="srun -n {ntasks} python {reader_path} {start.strftime("%Y-%m-%d")} {end.strftime("%Y-%m-%d")} {data_type} '
 
     if suffix:
         cmd += f'-s {suffix} '
@@ -57,7 +69,8 @@ def _make_sbatch_cmd(idx: int,
     return cmd
 
 
-def _make_parallel_cmd(start: datetime,
+def _make_parallel_cmd(reader_path: str,
+                       start: datetime,
                        end: datetime,
                        ntasks: int,
                        data_type: str,
@@ -65,7 +78,7 @@ def _make_parallel_cmd(start: datetime,
                        append: bool = True) -> str:
 
     if _is_slurm_available():
-        cmd = f'srun -n {ntasks} python {runner_path} {start.strftime("%Y-%m-%d")} {end.strftime("%Y-%m-%d")} {data_type} '
+        cmd = f'srun -n {ntasks} python {reader_path} {start.strftime("%Y-%m-%d")} {end.strftime("%Y-%m-%d")} {data_type} '
         if suffix:
             cmd += f'-s {suffix} '
 
@@ -73,7 +86,7 @@ def _make_parallel_cmd(start: datetime,
             cmd += '-a '
 
     else:
-        cmd = f'mpirun -n {ntasks} python {runner_path} {start.strftime("%Y-%m-%d")} {end.strftime("%Y-%m-%d")} {data_type} '
+        cmd = f'mpirun -n {ntasks} python {reader_path} {start.strftime("%Y-%m-%d")} {end.strftime("%Y-%m-%d")} {data_type} '
         if suffix:
             cmd += f'-s {suffix} '
 
@@ -83,13 +96,14 @@ def _make_parallel_cmd(start: datetime,
     return cmd
 
 
-def _make_serial_cmd(start: datetime,
+def _make_serial_cmd(reader_path: str,
+                     start: datetime,
                      end: datetime,
                      data_type: str,
                      suffix: str = None,
                      append=True) -> str:
 
-    cmd = f'python {runner_path} {start.strftime("%Y-%m-%d")} {end.strftime("%Y-%m-%d")} {data_type} '
+    cmd = f'python {reader_path} {start.strftime("%Y-%m-%d")} {end.strftime("%Y-%m-%d")} {data_type} '
 
     if suffix:
         cmd += f'-s {suffix} '
@@ -117,7 +131,8 @@ def _split_datetime_range(start: datetime, end: datetime, num_days: int) -> list
     return ranges
 
 
-def _batch_gen(start: datetime,
+def _batch_gen(reader_path: str,
+               start: datetime,
                end: datetime,
                max_days: int,
                ntasks: int,
@@ -126,12 +141,12 @@ def _batch_gen(start: datetime,
                append: bool = True,
                slurm_account: str = None) -> None:
     ranges = _split_datetime_range(start, end, max_days)
-
     cmds = []
     for idx, (start, end) in enumerate(ranges):
         if idx == 0:
             cmds.append(
-                _make_sbatch_cmd(idx,
+                _make_sbatch_cmd(reader_path,
+                                 idx,
                                  start,
                                  end,
                                  ntasks,
@@ -141,7 +156,8 @@ def _batch_gen(start: datetime,
                                  slurm_account=slurm_account))
         else:
             cmds.append(
-                _make_sbatch_cmd(idx,
+                _make_sbatch_cmd(reader_path,
+                                 idx,
                                  start,
                                  end,
                                  ntasks,
@@ -154,23 +170,25 @@ def _batch_gen(start: datetime,
     os.system(cmd)
 
 
-def _parallel_gen(start: datetime,
+def _parallel_gen(reader_path: str,
+                  start: datetime,
                   end: datetime,
                   ntasks: int,
                   gen_type: str,
                   suffix: str = None,
                   append: bool = True) -> None:
-    cmd = _make_parallel_cmd(start, end, ntasks, gen_type, suffix=suffix, append=append)
+    cmd = _make_parallel_cmd(reader_path, start, end, ntasks, gen_type, suffix=suffix, append=append)
     print(cmd)
     os.system(cmd)
 
 
-def _serial_gen(start: datetime,
+def _serial_gen(reader_path: str,
+                start: datetime,
                 end: datetime,
                 gen_type: str,
                 suffix: str = None,
                 append: bool = True) -> None:
-    cmd = _make_serial_cmd(start, end, gen_type, suffix=suffix, append=append)
+    cmd = _make_serial_cmd(reader_path, start, end, gen_type, suffix=suffix, append=append)
     print(cmd)
     os.system(cmd)
 
@@ -199,8 +217,10 @@ if __name__ == "__main__":
 
     def call_generator(gen_type):
         type_config = config.get_data_type(gen_type)
+        reader_path = get_reader_path(type_config.type)
         if args.batch:
-            _batch_gen(start_date,
+            _batch_gen(reader_path,
+                       start_date,
                        end_date,
                        type_config.batch_days,
                        type_config.num_tasks,
@@ -209,14 +229,16 @@ if __name__ == "__main__":
                        append=args.append,
                        slurm_account=args.slurm_account)
         elif args.parallel:
-            _parallel_gen(start_date,
+            _parallel_gen(reader_path,
+                          start_date,
                           end_date,
                           type_config.num_tasks,
                           gen_type,
                           suffix=args.suffix,
                           append=args.append)
         else:
-            _serial_gen(start_date,
+            _serial_gen(reader_path,
+                        start_date,
                         end_date,
                         gen_type,
                         suffix=args.suffix,
